refactor(diff_onedal): extract generic detectors into library-agnostic modules
PR-D of the 4-PR follow-up to #238 (oneDPL analysis).

Two detectors that were sitting in `abicheck/diff_onedal.py` are
naming-convention based and apply to any library that follows the
same convention, not just oneDAL. They are moved into dedicated
generic modules:

- `detect_serialization_tag_changes` (and its helpers
  `_collect_tag_constants`, `_looks_like_serialization_tag`, plus the
  `_TAG_SUFFIX_PATTERNS` / `_TAG_EXACT_LEAVES` constants) →
  `abicheck/diff_serialization.py`.
- `detect_missing_instantiations` (and `_looks_like_template_instantiation`
  + the equivalent of `_callable_stem`, which already lived in PR-B's
  `diff_templates` as `_strip_template_args`) → `abicheck/diff_templates.py`.

`abicheck/diff_onedal.py` now re-exports both at module top so
existing tests and external callers (`from abicheck.diff_onedal import …`)
continue to work unchanged. The `DetectOneDALPatterns` pipeline step
in `post_processing.py` imports directly from the new generic
locations.

The other oneDAL-shaped detectors (SYCL overload-set removal, CPU-
dispatch ISA drop, tag-type rename, default-template-arg change,
inline-body-references-renamed-member) are idiosyncratic enough that
leaving them in `diff_onedal.py` is the right call for now; further
extraction can happen incrementally as each one is generalised.

`_looks_like_template_instantiation` regex (`<[^<>]`) is preserved
exactly so `operator<` / `operator<<` / `operator<=>` do not
false-trigger the instantiation heuristic (regression test in
`test_onedal_detectors.py::TestTemplateInstantiationHeuristic`).

Test plan:
- pytest tests/ -m "not integration and not libabigail and not abicc
  and not slow and not golden" -q  →  6184 passed
- ruff check abicheck/ tests/  →  clean
- existing tests/test_onedal_detectors.py runs unchanged (78 tests)
